Starting commit
I'll write docs soon, but for now, there are lots of comments (yes, this
is unhelpful as a starting place).  Anyway, this is a Voxel engine in a
loose sense (Minecraft-style I guess) and it's roughly working.
Physics, such as they are, are a little wonky, and the way that block
textures are done could probably be improved.  But it's usable.  So
we'll call it v1.