Jumping bug fixed
Jumping is now (properly) unaffected by ground friction (as intended).

Reason: whether or not you're on the ground is calculated after position
is changed (sensibly) but friction affects all velocity when you're on
the ground.  So we need to update in the reverse natural order (position
first, then velocity, then acceleration).