Performance Improvements for Tall Chunks
Basically, a lot of things were being drawn that were never going to be
visible.

First, the very bottom of the chunk (which should never appear except in
a "fall through the world scenario," in which case I don't much care how
it looks).  This has been disabled directly.

Second, the edges of a chunk, which should be blocked, were still being
drawn because the chunk didn't know it didn't need them.  Fab.  So now
the chunk also contains its lateral boundaries.  This makes Map coding
slightly harder, as there is a contract that the block generation is
consistent, but the appropriate changes to the Demo Project should give
some recipes for making this easier (in the case of noise calculations,
for Demo 2, and in manual map geometry calculation, for Demo 1).

Also, StepUp has been turned back on (boo) because it makes wandering
around easier.  It will eventually be replaced by a slicker system, but
for now, here we are.

Also also, I compulsively fiddled with the RandomHelper class but I
didn't get anywhere.  That was mostly a waste of time.