True Voxels!
True voxels are, literally by definition, indivisible.  So I've removed
the ability for a block to have a boundingbox which does not fill up the
entire space.  This has other benefits elsewhere (e.g.: what if you want
to stack a full-height block on a half-height block?  It would have to
turn into two half-height blocks, requiring extra textures and more
complexity, etc.), and is not just a philosophical distinction.

This made several of the methods of BlockHandler no longer necessarily
abstract (such as the visual obstruction methods, or the boundingbox
methods).  Boundingboxes are now always unit cubes, and the visual
obstruction methods default to just returning IsVisible, though these
can be overridden in the case of a transparent (or partially
transparent) block.

Also, the "UseStep" flag has been moved around in the source and
switched back to false (though of course it's an Entity property, so can
be overridden, etc.  This is just flipping a default flag).  It's not so
pressing when the blocks are bigger.

Note that if you don't like big blocks, you can absolutely make them
smaller; just make the Entities' boundingboxes bigger, and in
particular, make your Avatar's height a little higher.  But be prepared
to see a (relative) drop in view distance.