Added some jump control
Two possible tags are added.

1) RetainFrictionInAir (default false).  This allows for moving around
in air as if you were on the ground.  Note that this remembers the
friction from the last ground you were on, so it's a little weird; if
you jump from grass, you do fine, if you jump from ice, you keep
sliding.  It's not a good mechanic but it was easy to add.  Default is
false.

2) UseIntendedVelocityInMidAir (default true).  This one is better;
basically your intended velocity is always moved toward (with decay
proportional to friction) on the ground, so with this tag, it also moves
toward it (with decay proportional to drag) in the air (or rather, at
all times, since you're always in the air).  Works good and helps people
not get stuck behind walls so bad!