Procedurally Generated Entities!
Entities can now be generated, along with everything else, in the
procedural generation method for Map.  They automatically spawn and
despawn properly, and these entities are managed completely separately
from manually-created entities (such as the PlayerAvatar), so
implementers won't have to worry about this system, or about fitting
their special stuff in it.

Note: this really brings to mind how really damn slow my physics engine
can be.  If you generate just a hundred or so mutually interactible
entities, you'll really see it (try turning the spawn radius up on Demo
Project 1).

I have some ideas to deal with this, but they're not in there yet.
Whatever it is, it will be intrinsic to the engine and not something
you'll have to deal with as an implementer.