Noninteractive Entities and Performance
There is now a flag for Entities which do not interact with other
Entities (e.g. for scenery) but which still want to use the physics
engine, for whatever reason.  These are separated at an efficient level,
so that their physics are a bit simpler to calculate.

Basic stress tests say that, without a drop in framerate (on my laptop,
at 60 fps), we can get about 40 mutually interacting objects, or a few
hundred non-interactive (but still physical) objects, or over a thousand
non-interactive non-physics objects.

Still not good enough!  But the lesson (at least for now) is to keep the
number of physics entities fairly low.