Rewrite of Map class and some etc.
Map was completely rewritten for clarity ("GetLookedAt" was unchanged
and needs some clarity help, but that aside...).  In particular, most of
the generated Entity management was moved to Cache, where it belongs.
The roles of the two classes are becoming more clear with time: Map
generates and examines, but Cache holds all the data that we generate.
And that's a good thing!

Now for some etc.:

1) Numerical was starting to hold some excess methods that weren't
really numerical.  I've moved them into a separate class called
"HelperMethods."  Little has been changed aside from that, but this
causes some cascading code changes, since several other classes use
those methods.

2) Demo projects have been updated to fit the new Map class.

3) Demo3 had a bugfix.  I thought it was a LookedAtBlock bug, but it was
just a translation error in the MouseOverBlock class.  Phew!