Chunks no longer contain neighbor information
This was a coding difficulty, and a huge (and hugely easy) source of
visual glitching, especially as we start allowing for Chunk editing.  So
Chunks no longer keep track of that.  Instead, they just query their Map
to get the neighbor data.

This does mean there are some "false positives" on visibility; the next
patch will fix that with appropriate re-querying.