Performance (updating neighbor information)
Now, when a chunk is loaded, it will automatically call its neighbors'
recalculate visual geometry data.  This process is fast enough that it's
not necessary to make it a separate task or anything.

Sidenote: Since IsReady doesn't call Request anymore, there isn't so
much need for caution when using it (mistake on my part, my bad;
accessors shouldn't have side-effects, even if the use case seems to be
super common).  So I got to eliminate a lot of extraneous uses of
IsInGridRange