BlockHandler and EntityBuilder are now singletons
For reals!  Like you can only make one of them or it will crash.  When
you make one (automatically done for you in the VoxelistGame class) it
will set that up as the single instance, and after that, the static
methods will work properly.  Of which there are many.

The singleton is a good way to have extensible static classes, which is
what I wanted from the start, but I only recently realized the trick of
hiding the singleton (instead of passing it around everywhere, which is
awkward).

Camera will probably get this same treatment.