Various efficiency improvements and a bugfix (?)
Slightly less data is being passed around for the physics engine, and
some checks are now "debug only."  The result is that "release mode"
runs a fair amount better.

Also, we have a Schrodenbug, which is thread race condition related and
extremely rare and difficult to replicate (but results in a crash).  I
think I fixed it, but I've said that before.